move class - MemberAuthProperty class to MemberProperties class
