+++ src/main/java/co/fineants/api/global/init/properties/MemberProperties.java
@@ -16,4 +16,22 @@ public class MemberProperties {
 	public MemberProperties(List<MemberAuthProperty> properties) {
 		this.properties = properties;
 	}
+
+	@Getter
+	public static class MemberAuthProperty {
+		private final String email;
+		private final String nickname;
+		private final String password;
+		private final String provider;
+		private final String roleName;
+
+		@ConstructorBinding
+		public MemberAuthProperty(String email, String nickname, String password, String provider, String roleName) {
+			this.email = email;
+			this.nickname = nickname;
+			this.password = password;
+			this.provider = provider;
+			this.roleName = roleName;
+		}
+	}
 }
